NOT WORKING BRANCH, just pushing logger package
