+++ cmd/mem_view/mem_view.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 	"syscall"
 
+	"github.com/AarC10/GSW-V2/lib/ipc"
 	"github.com/AarC10/GSW-V2/lib/tlm"
 	"github.com/AarC10/GSW-V2/lib/util"
-	"github.com/AarC10/GSW-V2/lib/ipc"
 	"github.com/AarC10/GSW-V2/proc"
 )
